Added Timer Settings and Refactored
The timer code works, but something in the pre-existing code for taking a picture messes up the timer.
