+++ PicIt/ContentView.swift
@@ -1,248 +1,31 @@
 //
-//  PicIt
-//
 
 import SwiftUI
-import Combine
-import AVFoundation
 
-final class CameraModel: ObservableObject {
-    private let service = CameraService()
-    
-    @Published var photo: Photo!
-    
-    @Published var showAlertError = false
-    
-    @Published var isFlashOn = false
-    
-    @Published var willCapturePhoto = false
-    
-    var alertError: AlertError!
-    
-    var session: AVCaptureSession
-    
-    private var subscriptions = Set<AnyCancellable>()
-    
-    init() {
-        self.session = service.session
-        
-        service.$photo.sink { [weak self] (photo) in
-            guard let pic = photo else { return }
-            self?.photo = pic
-        }
-        .store(in: &self.subscriptions)
-        
-        service.$shouldShowAlertView.sink { [weak self] (val) in
-            self?.alertError = self?.service.alertError
-            self?.showAlertError = val
-        }
-        .store(in: &self.subscriptions)
-        
-        service.$flashMode.sink { [weak self] (mode) in
-            self?.isFlashOn = mode == .on
-        }
-        .store(in: &self.subscriptions)
-        
-        service.$willCapturePhoto.sink { [weak self] (val) in
-            self?.willCapturePhoto = val
-        }
-        .store(in: &self.subscriptions)
-    }
-    
-    func configure() {
-        service.checkForPermissions()
-        service.configure()
-    }
-    
-    func capturePhoto() {
-        service.capturePhoto()
-    }
-    
-    func flipCamera() {
-        service.changeCamera()
-    }
-    
-    func zoom(with factor: CGFloat) {
-        service.set(zoom: factor)
-    }
-    
-    func switchFlash() {
-        service.flashMode = service.flashMode == .on ? .off : .on
-    }
-    
-    func photoTimer() {
 
-    }
-}
-
-struct CameraView: View {
-    @Environment(\.scenePhase) var scenePhase
-    
-    private static let timerDefault = 3
-    @StateObject var model = CameraModel()
-    
-    @State var currentZoomFactor: CGFloat = 1.0
-    
-    @State private var timeRemaining = Self.timerDefault
-    private var buttonText: String {
-        get {
-            let displayTime = timeRemaining < 0 ? "" : "\(timeRemaining)"
-            return "\(displayTime)"
-        }
-    }
-    @State private var buttonColor: Color = .white
-    let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
-    
-    var captureButton: some View {
-        Button(action: {
-            model.capturePhoto()
-        }, label: {
-            ZStack {
-                Circle()
-                    .foregroundColor(buttonColor)
-                    .frame(width: 80, height: 80, alignment: .center)
-                    .overlay(
-                        Circle()
-                            .stroke(Color.black.opacity(0.8), lineWidth: 2)
-                            .frame(width: 65, height: 65, alignment: .center)
-                    )
-                Text("\(buttonText)")
-            }
-            
-        })
-
-            .onReceive(timer, perform: { time in
-                if timeRemaining > 0 {
-                    buttonColor = .yellow
-                    timeRemaining -= 1
-                    if timeRemaining == 0 {
-                        model.capturePhoto()
-                        buttonColor = .green
-                    }
-                } else if timeRemaining == 0 {
-                    timeRemaining = -99
-                    buttonColor = .white
-                }
-
-            })
-            .onChange(of: scenePhase) { newPhase in
-              switch newPhase {
-              case .background:
-                timeRemaining = -99
-              case .active:
-                  timeRemaining = Self.timerDefault
-              case .inactive:
-                  timeRemaining = -99
-              @unknown default:
-                  timeRemaining = -99
-              }
-            }
+struct ContentView: View {
+    /*
+     We don't want to alter the value stored by AppStorage, so we should not access it directly, but only through the computed property. We could have used  "UserDefaults" instead of "AppStorage" to do this, but I prefer the conciseness of AppStorage.
+     Note: The default value (.delay.value) is only used if the setting doesn't exist (or at least I think so)
+     */
+    @AppStorage(PicItSetting.delay.key) private var _picitDelay: String = "\(PicItSetting.delay.value)"
+    private var picitDelay: Double {
+        return Double(_picitDelay) ?? 0.0
     }
     
-    var capturedPhotoThumbnail: some View {
-        Group {
-            if model.photo != nil {
-                Image(uiImage: model.photo.image!)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-                    .frame(width: 60, height: 60)
-                    .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                    // .animation(.spring())
-                
-            } else {
-                RoundedRectangle(cornerRadius: 10)
-                    .frame(width: 60, height: 60, alignment: .center)
-                    .foregroundColor(.black)
-            }
-        }
-    }
-    
-    var flipCameraButton: some View {
-        Button(action: {
-            model.flipCamera()
-        }, label: {
-            Circle()
-                .foregroundColor(Color.gray.opacity(0.2))
-                .frame(width: 45, height: 45, alignment: .center)
-                .overlay(
-                    Image(systemName: "camera.rotate.fill")
-                        .foregroundColor(.white))
-        })
+    var timer: PicItTimer {
+        let myTimer = PicItTimer(delay: picitDelay, interval: PicItDefault.interval, tolerance: PicItDefault.tolerance)
+        print("CREATED timer with: \(myTimer.delay), \(myTimer.interval), \(myTimer.tolerance)")
+        return myTimer
     }
     
     var body: some View {
-        GeometryReader { reader in
-            ZStack {
-                Color.black.edgesIgnoringSafeArea(.all)
-                
-                VStack {
-                    Button(action: {
-                        model.switchFlash()
-                    }, label: {
-                        Image(systemName: model.isFlashOn ? "bolt.fill" : "bolt.slash.fill")
-                            .font(.system(size: 20, weight: .medium, design: .default))
-                    })
-                    .accentColor(model.isFlashOn ? .yellow : .white)
-                    
-                    CameraPreview(session: model.session)
-                        .gesture(
-                            DragGesture().onChanged({ (val) in
-                                //  Only accept vertical drag
-                                if abs(val.translation.height) > abs(val.translation.width) {
-                                    //  Get the percentage of vertical screen space covered by drag
-                                    let percentage: CGFloat = -(val.translation.height / reader.size.height)
-                                    //  Calculate new zoom factor
-                                    let calc = currentZoomFactor + percentage
-                                    //  Limit zoom factor to a maximum of 5x and a minimum of 1x
-                                    let zoomFactor: CGFloat = min(max(calc, 1), 5)
-                                    //  Store the newly calculated zoom factor
-                                    currentZoomFactor = zoomFactor
-                                    //  Sets the zoom factor to the capture device session
-                                    model.zoom(with: zoomFactor)
-                                }
-                            })
-                        )
-                        .onAppear {
-                            model.configure()
-                        }
-                        .alert(isPresented: $model.showAlertError, content: {
-                            Alert(title: Text(model.alertError.title), message: Text(model.alertError.message), dismissButton: .default(Text(model.alertError.primaryButtonTitle), action: {
-                                model.alertError.primaryAction?()
-                            }))
-                        })
-                        .overlay(
-                            Group {
-                                if model.willCapturePhoto {
-                                    Color.black
-                                }
-                            }
-                        )
-                        .animation(.easeInOut)
-                    
-                    
-                    HStack {
-                        capturedPhotoThumbnail
-                        
-                        Spacer()
-                        
-                        captureButton
-                        
-                        Spacer()
-                        
-                        flipCameraButton
-                        
-                    }
-                    .padding(.horizontal, 20)
-                }
-        
-            }
-        }
+        CameraView(timer: timer)
     }
-        
 }
 
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
-        CameraView()
+        ContentView()
     }
 }
