+++ PicIt/PicItApp.swift
@@ -6,7 +6,7 @@ import SwiftUI
 struct PicItApp: App {
     var body: some Scene {
         WindowGroup {
-            CameraView()
+            ContentView()
         }
     }
 }
